requirements.txt
fastapi==0.112.0
uvicorn==0.32.0
biopython==1.83
scikit-learn==1.5.0
numpy==1.26.4
mlflow==2.16.0
requests==2.31.0
pydantic==2.9.0
fastapi
uvicorn[standard]
